refactor(plugins): rename create-objects capability modules to create-object
Consistent with other capability module names (create-object.ts). Also
renames the lazy-loaded exports from CreateObjects to CreateObject and
CreateObjectsNode to CreateObjectNode.
