+++ packages/plugins/plugin-assistant/src/capabilities/create-object.ts
