+++ packages/plugins/plugin-feed/src/FeedPlugin.tsx
@@ -6,7 +6,7 @@ import { ActivationEvent, Plugin } from '@dxos/app-framework';
 import { AppActivationEvents, AppPlugin } from '@dxos/app-toolkit';
 import { AttentionEvents } from '@dxos/plugin-attention/types';
 
-import { AppGraphBuilder, BlueprintDefinition, CreateObjects, OperationHandler, ReactSurface } from '#capabilities';
+import { AppGraphBuilder, BlueprintDefinition, CreateObject, OperationHandler, ReactSurface } from '#capabilities';
 import { meta } from '#meta';
 import { translations } from '#translations';
 import { Magazine, Subscription } from '#types';
@@ -17,7 +17,7 @@ export const FeedPlugin = Plugin.define(meta).pipe(
     activate: AppGraphBuilder,
   }),
   AppPlugin.addBlueprintDefinitionModule({ activate: BlueprintDefinition }),
-  AppPlugin.addCreateObjectModule({ activate: CreateObjects }),
+  AppPlugin.addCreateObjectModule({ activate: CreateObject }),
   AppPlugin.addOperationHandlerModule({ activate: OperationHandler }),
   AppPlugin.addSchemaModule({
     schema: [Subscription.Feed, Subscription.Post, Magazine.Magazine],
