+++ packages/plugins/plugin-inbox/src/InboxPlugin.node.ts
@@ -6,14 +6,14 @@ import { Plugin } from '@dxos/app-framework';
 import { AppPlugin } from '@dxos/app-toolkit';
 import { Event, Message } from '@dxos/types';
 
-import { CreateObjectsNode, OperationHandler } from '#capabilities';
+import { CreateObjectNode, OperationHandler } from '#capabilities';
 import { meta } from '#meta';
 import { Calendar, Mailbox } from '#types';
 
 // TODO(wittjosiah): Factor out shared modules.
 
 export const InboxPlugin = Plugin.define(meta).pipe(
-  AppPlugin.addCreateObjectModule({ activate: CreateObjectsNode }),
+  AppPlugin.addCreateObjectModule({ activate: CreateObjectNode }),
   AppPlugin.addOperationHandlerModule({ activate: OperationHandler }),
   AppPlugin.addSchemaModule({
     schema: [Event.Event, Mailbox.Mailbox, Calendar.Calendar, Message.Message],
