+++ packages/plugins/plugin-inbox/src/capabilities/create-object-node.ts
