+++ packages/plugins/plugin-outliner/src/capabilities/create-object.ts
