+++ packages/plugins/plugin-script/src/ScriptPlugin.node.ts
@@ -6,13 +6,13 @@ import { Plugin } from '@dxos/app-framework';
 import { AppPlugin } from '@dxos/app-toolkit';
 import { Script } from '@dxos/compute';
 
-import { AppGraphBuilder, BlueprintDefinition, CreateObjects, OperationHandler } from '#capabilities';
+import { AppGraphBuilder, BlueprintDefinition, CreateObject, OperationHandler } from '#capabilities';
 import { meta } from '#meta';
 
 export const ScriptPlugin = Plugin.define(meta).pipe(
   AppPlugin.addAppGraphModule({ activate: AppGraphBuilder }),
   AppPlugin.addBlueprintDefinitionModule({ activate: BlueprintDefinition }),
-  AppPlugin.addCreateObjectModule({ activate: CreateObjects }),
+  AppPlugin.addCreateObjectModule({ activate: CreateObject }),
   AppPlugin.addOperationHandlerModule({ activate: OperationHandler }),
   AppPlugin.addSchemaModule({ schema: [Script.Script] }),
   Plugin.make,
