+++ packages/plugins/plugin-script/src/capabilities/create-object.ts
