+++ packages/plugins/plugin-script/src/capabilities/index.ts
@@ -8,7 +8,7 @@ import type { Blueprint, OperationHandlerSet } from '@dxos/compute';
 
 export const AppGraphBuilder = Capability.lazy('AppGraphBuilder', () => import('./app-graph-builder'));
 export const BlueprintDefinition = Capability.lazy('BlueprintDefinition', () => import('./blueprint-definition'));
-export const CreateObjects = Capability.lazy('CreateObjects', () => import('./create-objects'));
+export const CreateObject = Capability.lazy('CreateObject', () => import('./create-object'));
 export const Compiler = Capability.lazy('Compiler', () => import('./compiler'));
 export const OperationHandler = Capability.lazy<OperationHandlerSet.OperationHandlerSet>(
   'OperationHandler',
