+++ packages/plugins/plugin-space/src/capabilities/index.ts
@@ -9,7 +9,7 @@ export * from './app-graph-builder';
 
 export const AppGraphSerializer = Capability.lazy('AppGraphSerializer', () => import('./app-graph-serializer'));
 export const CreateObject = Capability.lazy('CreateObject', () => import('./create-object'));
-export const CreateObjects = Capability.lazy('CreateObjects', () => import('./create-objects'));
+export const CreateObject = Capability.lazy('CreateObject', () => import('./create-object'));
 export const IdentityCreated = Capability.lazy('IdentityCreated', () => import('./identity-created'));
 export const Migrations = Capability.lazy('SpaceMigrations', () => import('./migrations'));
 export { NavigationHandler } from './navigation-handler';
